components/k8s-model-server/request-logger/requirements.txt
google-cloud-bigquery